refactor: extract AccessListDrawer sub-forms
Part of #103
